fix Logic All API

+++ ShoppingAPI/Controllers/ShoppingDeliveryAddressController.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ShoppingAPI.Common;
 using ShoppingAPI.Common.Models;
 using ShoppingAPI.Data.Models;
 using ShoppingAPI.Services.Interfaces;
@@ -11,14 +13,19 @@ namespace ShoppingAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    [Authorize(Roles = "Admin,SuperAdmin")]
+    [Authorize]
     public class ShoppingDeliveryAddressController : ControllerBase
     {
         private readonly IShoppingDeliveryAddressServices shoppingDeliveryAddressServices;
-        public ShoppingDeliveryAddressController(IShoppingDeliveryAddressServices shoppingDeliveryAddressServices)
+        private int UserId = -1;
+        private string roleName = "Guest";
+        public ShoppingDeliveryAddressController(IShoppingDeliveryAddressServices shoppingDeliveryAddressServices, IHttpContextAccessor httpContextAccessor)
         {
             this.shoppingDeliveryAddressServices = shoppingDeliveryAddressServices;
+            this.UserId = int.Parse(httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value);
+            this.roleName = httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.Role).Value;
         }
+
         [HttpGet]
         public async Task<IActionResult> ShoppingDeliveryAddresses()
         {
@@ -30,21 +37,28 @@ namespace ShoppingAPI.Controllers
                 Data = categories
             });
         }
-        [HttpGet("{id}"), AllowAnonymous]
+
+        //Get ShoppingDeliveryAddress from Id
+        [HttpGet("{id}")]
         public async Task<IActionResult> ShoppingDeliveryAddress(int id)
         {
             string roles = User.Claims.FirstOrDefault(x => x.Type.Equals(ClaimTypes.Role))?.Value ?? "";
             var UserId = User.Claims.FirstOrDefault(x => x.Type == ClaimTypes.NameIdentifier).Value;
+
             var shoppingDeliveryAddress = await shoppingDeliveryAddressServices.GetShoppingDeliveryAddressAsync(id);
+
+            //Check ShoppingDeliveryAddress Exists
             if (shoppingDeliveryAddress != null)
             {
-                if (roles.Equals("SuperAdmin") || roles.Equals("Admin") || UserId.Equals(shoppingDeliveryAddress.UserId.ToString()))
+                //check if admin or user for Address user
+                if (Library.isAdmin(roles) || UserId.Equals(shoppingDeliveryAddress.UserId.ToString()))
                     return Ok(new ResultApi
                     {
                         Status = (int)HttpStatusCode.OK,
                         Data = shoppingDeliveryAddress,
                         Success = true
                     });
+
                 return NotFound(new ResultApi
                 {
                     Status = (int)HttpStatusCode.NotFound,
@@ -55,11 +69,13 @@ namespace ShoppingAPI.Controllers
             return Unauthorized();
         }
 
+        //Insert ShoppingDelivery Address by User
         [HttpPost]
         public async Task<IActionResult> shoppingDeliveryAddress(ShoppingDeliveryAddress shoppingDelivery)
         {
             if (ModelState.IsValid)
             {
+                shoppingDelivery.UserId = this.UserId;
                 await shoppingDeliveryAddressServices.InsertShoppingDeliveryAddress(shoppingDelivery);
                 return Ok(new ResultApi
                 {
@@ -72,39 +88,51 @@ namespace ShoppingAPI.Controllers
             }
             return BadRequest();
         }
+        //Update Shopping DeliveryAddress by User
 
-        [HttpPut("InfomationUser")]
+        [HttpPut("ShoppingDeliveryAddress")]
         public async Task<IActionResult> PutInfomationUser(ShoppingDeliveryAddress shoppingDelivery)
         {
             if (ModelState.IsValid)
             {
-                var shoppingDeliveryAddressDb = await shoppingDeliveryAddressServices.GetShoppingDeliveryAddressAsync(shoppingDelivery.Id);
-
-                shoppingDeliveryAddressDb.PhoneNumber = shoppingDelivery.PhoneNumber;
-                shoppingDeliveryAddressDb.Address = shoppingDelivery.Address;
-
-                await shoppingDeliveryAddressServices.UpdateShoppingDeliveryAddress(shoppingDeliveryAddressDb);
-
-                return Ok(new ResultApi
+                if (this.UserId == shoppingDelivery.UserId)
                 {
-                    Status = (int)HttpStatusCode.OK,
-                    Success = true,
-                    Message = new[] { "Edit success" },
-                    Data = shoppingDeliveryAddressDb
-                });
+                    var shoppingDeliveryAddressDb = await shoppingDeliveryAddressServices.GetShoppingDeliveryAddressAsync(shoppingDelivery.Id);
+
+                    shoppingDeliveryAddressDb.PhoneNumber = shoppingDelivery.PhoneNumber;
+                    shoppingDeliveryAddressDb.Address = shoppingDelivery.Address;
+
+                    await shoppingDeliveryAddressServices.UpdateShoppingDeliveryAddress(shoppingDeliveryAddressDb);
+
+                    return Ok(new ResultApi
+                    {
+                        Status = (int)HttpStatusCode.OK,
+                        Success = true,
+                        Message = new[] { "Edit success" },
+                        Data = shoppingDeliveryAddressDb
+                    });
+                }
             }
-            return BadRequest();
+            return Unauthorized();
         }
 
-        [HttpDelete("InfomationUser")]
+        //Delete deleteShoppingDeliveryAddress
+
+        [HttpDelete("ShoppingDeliveryAddress")]
         public async Task<IActionResult> DeleteShoppingDeliveryAddress(int id)
         {
-            await shoppingDeliveryAddressServices.DeleteShoppingDeliveryAddress(id);
-            return Ok(new ResultApi
+            var shoppingbyUser = await shoppingDeliveryAddressServices.GetShoppingDeliveryAddressAsync(id);
+            //check if true user
+            if (shoppingbyUser.UserId == this.UserId)
             {
-                Success = true,
-                Message = new[] { "Delete Success" }
-            });
+                await shoppingDeliveryAddressServices.DeleteShoppingDeliveryAddress(id);
+                return Ok(new ResultApi
+                {
+                    Success = true,
+                    Message = new[] { "Delete Success" }
+                });
+            }
+            return Unauthorized();
         }
     }
 }
